feat: create new shared directory and move the components
