+++ src/shared/components/Badge/Badge.styles.ts
@@ -1,10 +1,10 @@
-import { theme, TButtonVariant, TButtonColor } from "../../theme";
+import { theme, TBadgeVariant, TButtonColor } from "../../../theme";
 
-export const getClassNames = (variant: TButtonVariant, color: TButtonColor) => {
+export const getClassNames = (variant: TBadgeVariant, color: TButtonColor) => {
   const defaultClassNames = Object.values(theme.defaultStyles).join(" ");
 
   const { textColor, border, backgroundColor } =
-    theme.variants.button[variant][color];
+    theme.variants.badge[variant][color];
 
   return `${defaultClassNames} ${textColor} ${border} ${backgroundColor}`;
 };
