+++ src/shared/components/Button/Buttons.styles.ts
@@ -1,10 +1,10 @@
-import { theme, TBadgeVariant, TButtonColor } from "../../theme";
+import { theme, TButtonVariant, TButtonColor } from "../../../theme";
 
-export const getClassNames = (variant: TBadgeVariant, color: TButtonColor) => {
+export const getClassNames = (variant: TButtonVariant, color: TButtonColor) => {
   const defaultClassNames = Object.values(theme.defaultStyles).join(" ");
 
   const { textColor, border, backgroundColor } =
-    theme.variants.badge[variant][color];
+    theme.variants.button[variant][color];
 
   return `${defaultClassNames} ${textColor} ${border} ${backgroundColor}`;
 };
